+++ packages/react/src/components/ErrorSummary/index.ts
@@ -1,4 +1,4 @@
-import { ErrorSummaryRoot } from './ErrorSummary';
+import { ErrorSummaryRoot } from './ErrorSummaryRoot';
 import { ErrorSummaryHeading } from './ErrorSummaryHeading';
 import { ErrorSummaryItem } from './ErrorSummaryItem';
 import ErrorSummaryList from './ErrorSummaryList';
@@ -22,7 +22,7 @@ ErrorSummary.Item.displayName = 'ErrorSummary.Item';
 ErrorSummary.Heading.displayName = 'ErrorSummary.Heading';
 ErrorSummary.List.displayName = 'ErrorSummary.List';
 
-export type { ErrorSummaryProps } from './ErrorSummary';
+export type { ErrorSummaryProps } from './ErrorSummaryRoot';
 export type { ErrorSummaryItemProps } from './ErrorSummaryItem';
 export type { ErrorSummaryHeadingProps } from './ErrorSummaryHeading';
 export type { ErrorSummaryListProps } from './ErrorSummaryList';
